Further Explanation
I was wondering a while why my scans of upgradeable decorations suddenly had a change in the filepath of the deco controller as I used a command block cause I set an anchor block in the command block. This should prevent people from having issues with command blocks on upgradeable decorations. I just copied the sentences over from the Scantool tutorial cause it gave me the right direction of what was happening. The Scantool isn't as bad to figure out yourself but making your custom style depends more on how the filesystem is set up (so you need to read at least the schematics tutorial).